requirements.txt
backcall==0.2.0
colorama==0.4.4
debugpy==1.5.0
decorator==5.1.0
entrypoints==0.3
h5py==3.4.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.8.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==301
pyzmq==22.3.0
six==1.16.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
dask==2.30.0
distributed==2.30.0
fsspec>=0.3.3
bokeh>=0.13.0
openpyxl
tqdm
graphviz
pyarrow
fastparquet